+++ src/Smuggler/Import.hs
@@ -1,97 +1,89 @@
-{-# OPTIONS_GHC -fno-warn-orphans #-}
+module Smuggler.Import where
 
-{-# LANGUAGE FlexibleInstances    #-}
+import Control.Monad (guard)
+import qualified Data.HashMap.Strict as HashMap ()
+import HsImpExp
+  ( IE (IEThingAbs, IEThingAll, IEThingWith, IEVar),
+    IEWrappedName (IEName),
+    ImportDecl (ideclHiding, ideclName),
+    LIE,
+    LIEWrappedName,
+  )
+import HsSyn (GhcRn)
+import Name (Name, nameSrcSpan)
+import PrelNames (pRELUDE_NAME)
+import RnNames (ImportDeclUsage)
+import SrcLoc
+  ( GenLocated (L),
+    SrcSpan (RealSrcSpan, UnhelpfulSpan),
+    srcSpanEndCol,
+    srcSpanStartCol,
+    srcSpanStartLine,
+    unLoc,
+  )
 
-module Smuggler.Import
-       ( Import (..)
-       , getLocationMap
-       ) where
+-- TODO: reuse more logic from GHC. Is it possible?
+unusedLocs :: ImportDeclUsage -> [(Int, Int)]
+unusedLocs (L (UnhelpfulSpan _) _, _, _) = []
+unusedLocs (L (RealSrcSpan loc) decl, used, unused)
+  | -- Do not remove `import M ()`
+    Just (False, L _ []) <- ideclHiding decl
+  = []
+  | -- Note [Do not warn about Prelude hiding]
+    -- TODO: add ability to support custom prelude
+    Just (True, L _ hides) <- ideclHiding decl
+  , not (null hides)
+  , pRELUDE_NAME == unLoc (ideclName decl)
+  = []
+  | {-
+        This is is not always correct, because instances may be imported
+        as in first case above
+    
+        -- Nothing used; drop entire decl
+        | null used = [ (lineNum, colNum)
+                      | lineNum <- [srcSpanStartLine loc .. srcSpanEndLine loc]
+                      , colNum <-  [srcSpanStartCol loc .. getEndColMax unused]
+                      ]
+    -}
 
-import BasicTypes (StringLiteral (..))
-import HsImpExp (IE (..), IEWildcard (..), IEWrappedName (..), ImportDecl (..), LIE, LIEWrappedName,
-                 LImportDecl)
-import HsSyn (HsModule (..))
-import Module (ModuleName, moduleName, moduleNameString)
-import Name (nameOccName)
-import OccName (occNameString)
-import RdrName (RdrName (..))
-import SrcLoc (GenLocated (L), Located, SrcSpan)
+    -- Everything imported is used; drop nothing
+    null unused
+  = []
+  | -- only part of non-hiding import is used
+    Just (False, L _ lies) <- ideclHiding decl
+  = unusedLies lies
+  | -- TODO: unused hidings
+    otherwise
+  = []
+ where
+  unusedLies :: [LIE GhcRn] -> [(Int, Int)]
+  unusedLies = concatMap lieToLoc
 
-import Smuggler.Loc (showL, showLoc)
+  lieToLoc :: LIE GhcRn -> [(Int, Int)]
+  lieToLoc (L _ lie) = case lie of
+    IEVar      _ name            -> lieNameToLoc name
+    IEThingAbs _ name            -> lieNameToLoc name
+    IEThingAll _ name            -> lieNameToLoc name
+    IEThingWith _ name _ names _ -> concatMap lieNameToLoc (name : names)
+    _                            -> []
 
-import qualified Data.HashMap.Strict as HashMap
-import qualified Text.Show as Show (show)
+  lieNameToLoc :: LIEWrappedName Name -> [(Int, Int)]
+  lieNameToLoc lieName = do
+    L _ (IEName (L (RealSrcSpan lieLoc) name)) <- [lieName]
+    guard $ name `elem` unused
+    pure (srcSpanStartLine lieLoc, srcSpanStartCol lieLoc)
 
-data Import name = Import
-    { impName      :: Located ModuleName -- ^ Module name.
-    , impPkgQual   :: Maybe StringLiteral  -- ^ Package qualifier.
-    , impQualified :: Bool          -- ^ True => qualified
-    , impImplicit  :: Bool          -- ^ True => implicit import (of Prelude)
-    , impAs        :: Maybe (Located ModuleName)  -- ^ as Module
-    , impHiding    :: Maybe (Bool, Located [LIE name]) -- ^ (True => hiding, names)
-    , impId        :: Int -- ^ For hashing
-    } deriving (Eq, Generic)
+  getEndColMax :: [Name] -> Int
+  getEndColMax u = listMax $ map (findColLoc . nameSrcSpan) u
 
-instance (Show name) => Show (Import name) where
-    show i = let L _ md = impName i in "Module: " ++ moduleNameString md ++
-        case impHiding i of
-            Just (b, L _ lies) -> " " ++ show b ++ " " ++ showLIEs lies
-            Nothing            -> ""
+  findColLoc :: SrcSpan -> Int
+  findColLoc (RealSrcSpan   l) = srcSpanEndCol l
+  findColLoc (UnhelpfulSpan _) = defaultCol
 
-showLIEs :: (Show name) => [LIE name] -> String
-showLIEs []   = ""
-showLIEs lies = "[ " ++ intercalate ", " (map showLIE lies) ++"]"
+defaultCol :: Int
+defaultCol = 120
 
-showLIE :: forall name . (Show name) => LIE name -> String
-showLIE (L _ ie) = showIE ie
-  where
-    showIE :: IE name -> String
-    showIE (IEVar lieWrappedName)       = showL " IEVar " showIEWrappedName lieWrappedName
-    showIE (IEThingAbs lieWrappedName)  = showL " IEThingAbs " showIEWrappedName lieWrappedName
-    showIE (IEThingAll lieWrappedName)  = showL " IEThingAll " showIEWrappedName lieWrappedName
-    showIE (IEThingWith lieWrappedName wildCard lieWNs _) = showL " IEThingWith " showIEWrappedName lieWrappedName
-        ++ showIEWildCard wildCard
-        ++ "[" ++ intercalate ", " (map showLIEWrappedName lieWNs) ++ "]"
-    showIE _ = "Not going to be"
-
-    showLIEWrappedName :: LIEWrappedName name -> String
-    showLIEWrappedName (L l ieWrappedName) = showLoc l ++ showIEWrappedName ieWrappedName
-
-    showIEWrappedName :: IEWrappedName name -> String
-    showIEWrappedName (IEName located)    = showL " IEName " show located
-    showIEWrappedName (IEPattern located) = showL " IEPattern " show located
-    showIEWrappedName (IEType located)    = showL " IEType " show located
-
-
-    showIEWildCard :: IEWildcard -> String
-    showIEWildCard NoIEWildcard   = ""
-    showIEWildCard (IEWildcard n) = "(..)" ++ show n
-
-instance Show RdrName where
-    show (Unqual occName)       = occNameString occName
-    show (Qual modName occName) = moduleNameString modName ++ " " ++ occNameString occName
-    show (Orig md occName)      = moduleNameString (moduleName md) ++ " " ++ occNameString occName
-    show (Exact nm)             = occNameString $ nameOccName nm
-
-instance Hashable (Import name) where
-    hashWithSalt salt = hashWithSalt salt . impId
-
-getLocationMap :: (Eq name) => HsModule name -> HashMap (Import name) SrcSpan
-getLocationMap HsModule{..} =  HashMap.fromList $ fromDecls hsmodImports
-  where
-    fromDecls :: [LImportDecl name] -> [(Import name, SrcSpan)]
-    fromDecls lidecls =  zipWith (curry toImportPairs) lidecls [0 .. ]
-
-    toImportPairs :: (LImportDecl name, Int) -> (Import name, SrcSpan)
-    toImportPairs (L l impDecl, i) = (toMyImport impDecl i, l)
-
-    toMyImport :: ImportDecl name -> Int -> Import name
-    toMyImport ImportDecl{..} i = Import
-        { impName      = ideclName
-        , impPkgQual   = ideclPkgQual
-        , impQualified = ideclQualified
-        , impImplicit  = ideclImplicit
-        , impAs        = ideclAs
-        , impHiding    = ideclHiding
-        , impId        = i
-        }
+listMax :: [Int] -> Int
+listMax []           = defaultCol
+listMax [x         ] = x
+listMax (x : y : xs) = listMax ((if x >= y then x else y) : xs)
